gas: change debt entry amount to uint96 to pack in one slot
should be enough even with 18 decimal stablecoins
uint96.max / 10^18 = ~79*10^9

+++ src/libs/DebtQueue.sol
@@ -23,7 +23,7 @@ library DebtQueue {
 
     struct DebtEntry {
         address account;
-        uint256 amount;
+        uint96 amount;
     }
 
     /**
